Create a new directory for scripts
