+++ views/template.mustache
@@ -30,6 +30,6 @@
 	<h1>{{pageTitle}}</h1>
 	<p>Baseline: {{totalItems}}</p>
 	{{{tableSections}}} <!-- Use triple mustache for raw HTML content -->
- <script src="../index.js"></script>
+ <script src="../scripts/collapsible.js"></script>
 </body>
 </html>
